refactor: strUpController update fn
