+++ roulette/models/fieldAndTrackModel.js
@@ -8,6 +8,7 @@ import { rouletteSeries } from './consts/rouletteSeries.js';
 // import { fullBets } from "./functions/fullBets.js";
 // import { rltPos } from "./consts/rltPos.js";
 import { chipCountAll } from "./functions/strUp/chipCountAll.js";
+import { strUpCalculations } from "./functions/strUp/strUpCalculations.js";
 
 // выбор минимума максимума рулетки:
 document.getElementById('minmax').addEventListener('change', function () {
@@ -28,6 +29,8 @@ fillArr = chipCountAll("streets", 5);
 fillArr = chipCountAll("corners", 5);
 fillArr = chipCountAll("six_lines", 5);
 console.log(fillArr);
+const pay = fillArr.reduce((acc, number) => acc + number, 0);
+console.log(`выплата: ${pay}`);
 
 // пересекается ли трек с полем:
 export const trackOverField = true;
